Move `PyramidalDataset` and `PyramidalBdv` classes to `pyramid.fiji` package
